Moves Enigma methods to LetterShift and LetterShiftTest
