+++ encrypted.txt
@@ -1,1 +1,1 @@
-{:encryption=>"bsknorycffdkxgkcnskpoep", :key=>"80590", :date=>"230219"}
+{:encryption=>"sosqenffwblnncsfdosseax", :key=>"70139", :date=>"230219"}
