+++ lib/enigma.rb
@@ -1,6 +1,8 @@
 require "./lib/shifts"
+require "./lib/letter_shift"
 
 class Enigma
+  include LetterShift
   attr_reader :message,
               :key,
               :offset,
@@ -22,13 +24,14 @@ class Enigma
   end
 
   def create_shifted_arrays(key,offset)
-    if key != nil && offset != nil
-      # binding.pry
-      @a_array = @alphabet.rotate(@shifts.shifts[:a])
-      @b_array = @alphabet.rotate(@shifts.shifts[:b])
-      @c_array = @alphabet.rotate(@shifts.shifts[:c])
-      @d_array = @alphabet.rotate(@shifts.shifts[:d])
+    while key == nil && offset == nil
+      # key = @shifts.random_key
+      # offset = @shifts.today_offset
     end
+    @a_array = @alphabet.rotate(@shifts.shifts[:a])
+    @b_array = @alphabet.rotate(@shifts.shifts[:b])
+    @c_array = @alphabet.rotate(@shifts.shifts[:c])
+    @d_array = @alphabet.rotate(@shifts.shifts[:d])
      #could use a while loop here
     # if @offset = nil
     # end
@@ -36,57 +39,6 @@ class Enigma
 
   def char_index
     @message.chars.map {|char| @alphabet.index(char)}
-  end
-
-  def shift_a
-    count = 0
-    char_index.each_with_index do |char,index|
-      if index == count
-        @message[index] = @a_array[char]
-        count += 4
-      end
-    end
-    @message
-  end
-
-  def shift_b
-    count = 1
-    char_index.each_with_index do |char,index|
-      if index == count
-        @message[index] = @b_array[char]
-        count += 4
-      end
-    end
-    @message
-  end
-
-  def shift_c
-    count = 2
-    char_index.each_with_index do |char,index|
-      if index == count
-        @message[index] = @c_array[char]
-        count += 4
-      end
-    end
-    @message
-  end
-
-  def shift_d
-    count = 3
-    char_index.each_with_index do |char,index|
-      if index == count
-        @message[index] = @d_array[char]
-        count += 4
-      end
-    end
-    @message
-  end
-
-  def shift_all
-    shift_a
-    shift_b
-    shift_c
-    shift_d
   end
 
   def encrypt_hash(encryption,key,offset)
